requirements.txt
asgiref==3.5.0
coverage==6.3.1
Django==3.2.12
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
